+++ idea/tests/org/jetbrains/kotlin/idea/codeInsight/OverrideImplementTest.kt
@@ -248,6 +248,14 @@ class OverrideImplementTest : AbstractOverrideImplementTest() {
         doOverrideDirectoryTest("foo")
     }
 
+    fun testNullableJavaType() {
+        doOverrideDirectoryTest("foo")
+    }
+
+    fun testNullableKotlinType() {
+        doOverrideDirectoryTest("foo")
+    }
+
     fun testAbstractAndNonAbstractInheritedFromInterface() {
         doImplementFileTest("getFoo")
     }